Add aidtransparency homepage sanity checks
This adds some sanity checks as to the length of content on the aidtransparency
homepage. It does this by ensuring that the title and content of the news article
summaries are not really short (nothing there) and not really long (an entire
article is being displayed).

Note: The specified min and max lengths are based on observation at the time of
creation, and may need modifying at a later point, for example if an article
with a 1-word title is created.