Add a test for minimum response size
This adds a test on all responses to ensure that the content size is
greater than a set size. This will then detect cases where a 200 status
code is returned, but there is no (or very litte) content.

The limit of 4000 is set by looking at the expected responses, finding
that the smallest is around 4700, and providing a bit of scope for
slightly smaller responses.

An improved version of this test would provide a default value, then
allow for customisation on a per-request basis.