Add additional datastore tests
The datastore test_recent_activities() test is failing between midnight
and 6am on a daily basis.

This adds a pair of additional tests to check for activities further in
the past. These will demonstrate whether the issue is simply related to
activities taking more than a day to update, or whether there is
something more major happening.