Fix datastore morning errors
This should fix the datastore test that fails between midnight and
around 06:15 UTC every morning.

The test is failing because the datastore takes around 30 hours to
update with the new data. This adds a check so that there is extra
leniency beyond the first 24 hours in which the datastore may
continue to update data.

As more data is added, the extra few hours may not be enough,
causing a need to either extend the period or change how the
datastore updates itself with new activities.