Further increase Dashboard tolerance
Changes the number of days we allow the Dashboard to not generate
from 7 to 10. Current failures appear due to low disk space, which has
now been increased. We expect the Dashboard to complete its next
regeneration on Saturday. From this point, the tolerance can be
reduced back to the original 2 days.