Update test_dashboard.py
Test a page that is generated late in the process.

The chosen page is the page that, during the last generation of the Dashboard, was the last page to generate that will consistently appear. One other page was generated later, though this refers to a custom namespace used by one organisation, so is not a good candidate to rely upon.