Increase the allowed difference for the number of publishers
Increases the allowed difference between the number of publishers reported by the Registry and the Dashboard. This is necessary as the Dashboard now only updates once every three days: https://discuss.iatistandard.org/t/changes-to-the-updating-of-dashboard-publisher-statistics/1172

Therefore there is a higher chance that the number of publishers will be inconsistent as there will be more time for new publishers to be created on the Registry.