Increase min value floors
There are a number of constants in place to check that a notable amount of data does not suddenly disappear. These values were set 15 months ago.

In the past 15 months, each of the actual values has increased by 30-50%.

This change increases the floor values so that they will be able to catch smaller errors should something go wrong.