Mark Dashboard publisher count consistency as xfail
The number of publishers on the Registry is increasing rapidly as Belgian NGOs are signing up in advance of a Belgian government publishing deadline. As such the frequency of the Dashboard generation process cannot keep up. Marking as xfail as the number of NGO registrations is expected to tail off as the Belgian government publishing deadline passes.