Revert Dashboard xfail tests
The production Dashboard has completed regeneration process (as of this lunchtime). Therefore these tests will now pass and xfail marks have been removed.

Despite server debugging, it is not known at this time why the Dashboard did not regenerate for so long. It is also unknown why the Dashboard regenerated with no configuration or code changes.